backend/requirements/base.txt
# Core
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-environ>=0.11,<1.0
django-cors-headers>=4.3,<5.0
django-filter>=24.0,<25.0

# Auth
djangorestframework-simplejwt>=5.3,<6.0

# Database
psycopg2-binary>=2.9,<3.0

# Tasks
celery>=5.3,<6.0
redis>=5.0,<6.0

# API docs
drf-spectacular>=0.27,<1.0

# Image handling
Pillow>=10.0,<11.0

# Utilities
python-slugify>=8.0,<9.0
